requirements.txt
flask
Flask-SQLAlchemy
mysqlclient
alembic
pytest
flask-jwt-extended